backend/requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
bcrypt
beautifulsoup4
blinker
bs4
cachetools
certifi
charset-normalizer
chardet
click
dataclasses-json
deep-translator
distro
fastapi
faiss-cpu
fpdf
frozenlist
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
google-genai
googleapis-common-protos
greenlet
grpcio
grpcio-status
gtts
h11
httpcore
httpx
idna
importlib-metadata
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain_google_genai
langsmith
markdown2
MarkupSafe
marshmallow
matplotlib
mdurl
mistune
multidict
mypy-extensions
netlas
numpy
openpyxl
packaging
pandas
passlib
pefile
pexels-api-py
Pillow
plotly
proto-plus
protobuf
pyarrow
pyasn1
pyasn1-modules
pydeck
pydantic
pydantic_core
pyaudio
pydub
pyjwt
PyPDF2
Pygments
python-dateutil
python-dotenv
python-jose
python-multipart
pytz
PyYAML
referencing
requests
rich
rpds-py
rsa
scapy
seaborn
setuptools
six
smmap
sniffio
SpeechRecognition
SQLAlchemy
tenacity
toml
toolz
tornado
tqdm
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
validators
watchdog
whois
wheel
yarl
youtube-transcript-api
zipp
dnspython
